+++ exonum/src/node/consensus.rs
@@ -511,8 +511,6 @@ impl NodeHandler {
 
         metric!("node.mempool", pool_len);
 
-        println!("commited {}", committed_txs);
-
         let height = self.state.height();
         info!(
             "COMMIT ====== height={}, proposer={}, round={}, committed={}, pool={}, hash={}",
@@ -775,9 +773,9 @@ impl NodeHandler {
                     self.sign_message(TransactionsRequest::new(&peer, &txs))
                         .into()
                 }
-                RequestData::PoolTransactions => self
-                    .sign_message(PoolTransactionsRequest::new(&peer))
-                    .into(),
+                RequestData::PoolTransactions => {
+                    self.sign_message(PoolTransactionsRequest::new(peer)).into()
+                }
                 RequestData::BlockTransactions => {
                     let txs: Vec<_> = match self.state.incomplete_block() {
                         Some(incomplete_block) => {